% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pshift.R
\name{pshift}
\alias{pshift}
\title{Compute distribution function of a shifted distribution}
\usage{
pshift(dist)
}
\arguments{
\item{dist}{character string indicating distribution name. It can be set as:
\itemize{
  \item beta Beta Distribution
  \item binom Binomial Distribution
  \item cauchy Cauchy Distribution
  \item chisq Chi-Square Distribution
  \item exp Exponential Distribution
  \item f F Distribution
  \item gamma Gamma Distribution
  \item geom Geometric Distribution
  \item hyper Hypergeometric Distribution
  \item logis Logistic Distribution
  \item lnorm Log Normal Distribution
  \item nbinom Negative Binomial Distribution
  \item norm Normal Distribution
  \item pois Poisson Distribution
  \item t Student t Distribution
  \item unif Uniform Distribution
  \item weibull Weibull Distribution
  \item wilcox Wilcoxon Rank Sum Statistic Distribution
  \item signrank Wilcoxon Signed Rank Statistic Distribution
}}
}
\value{
numeric vector of the distribution function of a specified distribution
}
\description{
Compute distribution function of a shifted distribution
}
\examples{
# Poisson Distribution
pp <- pshift("pois")
zero_shift <- pp(q = c(1,2,3,4,5), lambda = 1.5, shift = 0)
zero_shift
one_shift <- pp(q = c(1,2,3,4,5), lambda = 1.5, shift = 1)
one_shift
two_shift <- pp(q = c(1,2,3,4,5), lambda = 1.5, shift = 2)
two_shift

# Student t distribution
pt <- pshift("t")
zero_shift <- pt(q = c(1,2,3,4,5), df = 40, shift = 0)
zero_shift
one_shift <- pt(q = c(1,2,3,4,5), df = 40, shift = 1)
one_shift
two_shift <- pt(q = c(1,2,3,4,5), df = 40, shift = 2)
two_shift

}

