% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qshift.R
\name{qshift}
\alias{qshift}
\title{Compute quantile function of a shifted distribution}
\usage{
qshift(dist)
}
\arguments{
\item{dist}{dist character string indicating distribution name. It can be set as:
\itemize{
  \item beta Beta Distribution
  \item binom Binomial Distribution
  \item cauchy Cauchy Distribution
  \item chisq Chi-Square Distribution
  \item exp Exponential Distribution
  \item f F Distribution
  \item gamma Gamma Distribution
  \item geom Geometric Distribution
  \item hyper Hypergeometric Distribution
  \item logis Logistic Distribution
  \item lnorm Log Normal Distribution
  \item nbinom Negative Binomial Distribution
  \item norm Normal Distribution
  \item pois Poisson Distribution
  \item t Student t Distribution
  \item unif Uniform Distribution
  \item weibull Weibull Distribution
  \item wilcox Wilcoxon Rank Sum Statistic Distribution
  \item signrank Wilcoxon Signed Rank Statistic Distribution
}}
}
\value{
numeric vector of the quantile function of a specified distribution
}
\description{
Compute quantile function of a shifted distribution
}
\examples{

# Negative Binomial distribution
qnb <- qshift("nbinom")
zero_shift <- qnb(p = c(0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1), size = 6, mu = 4, shift = 0)
zero_shift
one_shift <- qnb(p = c(0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1), size = 6, mu = 4, shift = 1)
one_shift
two_shift <- qnb(p = c(0,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1), size = 6, mu = 4, shift = 2)
two_shift

# Chi Squared Distribution
qc <- qshift("chisq")
zero_shift <- qc(p = c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1), df = 2, shift = 0)
zero_shift
one_shift <- qc(p = c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1), df = 2, shift = 1)
one_shift
two_shift <- qc(p = c(0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1), df = 2, shift = 2)
two_shift

}

