% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshift.R
\name{rshift}
\alias{rshift}
\title{Perform Random generation for a shifted distribution}
\usage{
rshift(dist)
}
\arguments{
\item{dist}{character string indicating distribution name. It can be set as:
\itemize{
  \item beta Beta Distribution
  \item binom Binomial Distribution
  \item cauchy Cauchy Distribution
  \item chisq Chi-Square Distribution
  \item exp Exponential Distribution
  \item f F Distribution
  \item gamma Gamma Distribution
  \item geom Geometric Distribution
  \item hyper Hypergeometric Distribution
  \item logis Logistic Distribution
  \item lnorm Log Normal Distribution
  \item nbinom Negative Binomial Distribution
  \item norm Normal Distribution
  \item pois Poisson Distribution
  \item t Student t Distribution
  \item unif Uniform Distribution
  \item weibull Weibull Distribution
  \item wilcox Wilcoxon Rank Sum Statistic Distribution
  \item signrank Wilcoxon Signed Rank Statistic Distribution
}}
}
\value{
numeric vector of random generated values from a specified distribution
}
\description{
Perform Random generation for a shifted distribution
}
\examples{

# Binomial Distribution
rb <- rshift("binom")
set.seed(1)
zero_shift <- rb(n = 10000, size = 6, prob = 0.5, shift = 0)
zero_shift
set.seed(1)
one_shift <- rb(n = 10000, size = 6, prob = 0.5, shift = 1)
one_shift
set.seed(1)
two_shift <- rb(n = 10000, size = 6, prob = 0.5, shift = 2)
two_shift

# Weibull Distribution
rw <- rshift("weibull")
set.seed(1)
zero_shift <- rw(n = 1000, shape = 1.4, scale = 3, shift = 0)
zero_shift
set.seed(1)
one_shift <- rw(n = 1000, shape = 1.4, scale = 3, shift = 1)
one_shift
set.seed(1)
two_shift <- rw(n = 1000, shape = 1.4, scale = 3, shift = 2)
two_shift

}

